+++ src/command/mod.rs
@@ -144,6 +144,7 @@ pub mod exec; // command to exec into pods
 pub mod namespaces; // commands relating to namespaces
 pub mod nodes; // commands relating to nodes
 pub mod pods; //commands relating to pods
+pub mod portforwards; // commands for forwarding ports
 pub mod statefulsets; // commands for statefulsets
 pub mod volumes; // commands relating to volumes
